feat: add sortable table hook
